tgalerts/requirements.txt
Telethon==1.40.0
python-dotenv==1.1.0
mysql-connector-python==8.4.0